BrianRot/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
gtts
moviepy
requests
python-multipart
fastapi-cors
